+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.7.4 (upcoming)
 
 - Add `disabledRules` and `enabledRules` settings (requires update of language extensions)
+- Add `disable rule` quick fix
 - Fix a bug where the `codeAction` request gets stuck in infinite loop
 - Fix another NullPointerException for word2vec